+++ glean-core/src/metrics/experiment.rs
@@ -3,7 +3,7 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use serde::{Deserialize, Serialize};
-use serde_json::json;
+use serde_json::{json, Map as JsonMap, Value as JsonValue};
 use std::collections::HashMap;
 
 use crate::error_recording::{record_error, ErrorType};
@@ -35,6 +35,21 @@ const MAX_EXPERIMENTS_EXTRAS_SIZE: usize = 20;
 pub struct RecordedExperimentData {
     pub branch: String,
     pub extra: Option<HashMap<String, String>>,
+}
+
+impl RecordedExperimentData {
+    // For JSON, we don't want to include {"extra": null} -- we just want to skip
+    // extra entirely. Unfortunately, we can't use a serde field annotation for this,
+    // since that would break bincode serialization, which doesn't support skipping
+    // fields. Therefore, we use a custom serialization function just for JSON here.
+    pub fn as_json(&self) -> JsonValue {
+        let mut value = JsonMap::new();
+        value.insert("branch".to_string(), json!(self.branch));
+        if self.extra.is_some() {
+            value.insert("extra".to_string(), json!(self.extra));
+        }
+        JsonValue::Object(value)
+    }
 }
 
 /// An experiment metric.