+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     in the baseline ping ([`glean.baseline.locale`](https://github.com/mozilla/glean/blob/c261205d6e84d2ab39c50003a8ffc3bd2b763768/glean-core/metrics.yaml#L28-L42))
     is redundant and will be removed by the end of the quarter.
   * Drop the Glean handle and move state into glean-core ([#664](https://github.com/mozilla/glean/pull/664))
+  * If an experiment includes no `extra` fields, it will no longer include `{"extra": null}` in the JSON paylod.
 * Android:
   * Collections performed before initialization (preinit tasks) are now dispatched off
     the main thread during initialization.